requirements.txt
filelock==3.14.0
fsspec==2024.5.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.2
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sympy==1.12
torch==2.3.0
torchvision==0.18.0
typing_extensions==4.12.0
tzdata==2024.1
